style: implement theme toggle functionality in NavBar and update ThemeIcon component
